+++ dst/controllers/serviceType.js
@@ -39,11 +39,11 @@ function add(req, res) {
                 try {
                     req.body.id = '';
                     let serviceType = createFromBody(req);
-                    const { totalCount } = yield serviceTypeService.search({
+                    const { data } = yield serviceTypeService.search({
                         project: { ids: [req.project.id] },
                         identifiers: [serviceType.identifier]
                     });
-                    if (totalCount > 0) {
+                    if (data.length > 0) {
                         throw new Error('既に存在する興行区分コードです');
                     }
                     serviceType = yield serviceTypeService.create(serviceType);
@@ -75,9 +75,11 @@ function getList(req, res) {
                 endpoint: process.env.API_ENDPOINT,
                 auth: req.user.authClient
             });
-            const result = yield serviceTypeService.search({
-                limit: req.query.limit,
-                page: req.query.page,
+            const limit = Number(req.query.limit);
+            const page = Number(req.query.page);
+            const { data } = yield serviceTypeService.search({
+                limit: limit,
+                page: page,
                 project: { ids: [req.project.id] },
                 ids: [req.query.id],
                 name: req.query.name,
@@ -85,8 +87,10 @@ function getList(req, res) {
             });
             res.json({
                 success: true,
-                count: result.totalCount,
-                results: result.data
+                count: (data.length === Number(limit))
+                    ? (Number(page) * Number(limit)) + 1
+                    : ((Number(page) - 1) * Number(limit)) + Number(data.length),
+                results: data
             });
         }
         catch (err) {