+++ src/controllers/distributions.ts
@@ -35,10 +35,10 @@ export async function add(req: Request, res: Response): Promise<void> {
                     id: req.body.id,
                     name: req.body.name
                 };
-                const searchDistribution = await distributionService.searchDistribution({
+                const { data } = await distributionService.searchDistribution({
                     id: req.body.id
                 });
-                if (searchDistribution.totalCount > 0) {
+                if (data.length > 0) {
                     message = '配給コードが既に登録されています。';
                 } else {
                     await distributionService.createDistribution(distribution);
@@ -72,16 +72,20 @@ export async function getList(req: Request, res: Response): Promise<void> {
             auth: req.user.authClient
         });
 
-        const result = await distributionService.searchDistribution({
-            limit: req.query.limit,
-            page: req.query.page,
+        const limit = Number(req.query.limit);
+        const page = Number(req.query.page);
+        const { data } = await distributionService.searchDistribution({
+            limit: limit,
+            page: page,
             id: req.query.id,
             name: req.query.name
         });
         res.json({
             success: true,
-            count: result.totalCount,
-            results: result.data.map((t) => {
+            count: (data.length === Number(limit))
+                ? (Number(page) * Number(limit)) + 1
+                : ((Number(page) - 1) * Number(limit)) + Number(data.length),
+            results: data.map((t) => {
                 return {
                     id: t.id,
                     name: t.name