+++ src/controllers/ticketType.ts
@@ -46,11 +46,11 @@ export async function add(req: Request, res: Response): Promise<void> {
                 let ticketType = createFromBody(req);
 
                 // 券種コード重複確認
-                const { totalCount } = await offerService.searchTicketTypes({
+                const { data } = await offerService.searchTicketTypes({
                     project: { ids: [req.project.id] },
                     identifier: `^${ticketType.identifier}$`
                 });
-                if (totalCount > 0) {
+                if (data.length > 0) {
                     throw new Error(`既に存在する券種コードです: ${ticketType.identifier}`);
                 }
 
@@ -275,9 +275,11 @@ export async function getList(req: Request, res: Response): Promise<void> {
             }
         }
 
-        const result = await offerService.searchTicketTypes({
-            limit: req.query.limit,
-            page: req.query.page,
+        const limit = Number(req.query.limit);
+        const page = Number(req.query.page);
+        const { data } = await offerService.searchTicketTypes({
+            limit: limit,
+            page: page,
             project: { ids: [req.project.id] },
             ids: (ticketTypeIds.length > 0) ? ticketTypeIds : undefined,
             identifier: req.query.identifier,
@@ -285,8 +287,10 @@ export async function getList(req: Request, res: Response): Promise<void> {
         });
         res.json({
             success: true,
-            count: result.totalCount,
-            results: result.data.map((t) => {
+            count: (data.length === Number(limit))
+                ? (Number(page) * Number(limit)) + 1
+                : ((Number(page) - 1) * Number(limit)) + Number(data.length),
+            results: data.map((t) => {
                 return {
                     ...t,
                     ticketCode: t.identifier
@@ -329,14 +333,20 @@ export async function getTicketTypeGroupList(req: Request, res: Response): Promi
             endpoint: <string>process.env.API_ENDPOINT,
             auth: req.user.authClient
         });
-        const { totalCount, data } = await offerService.searchTicketTypeGroups({
-            limit: 100,
+
+        const limit = 100;
+        const page = 1;
+        const { data } = await offerService.searchTicketTypeGroups({
+            limit: limit,
+            page: page,
             project: { ids: [req.project.id] },
             ticketTypes: [req.params.ticketTypeId]
         });
         res.json({
             success: true,
-            count: totalCount,
+            count: (data.length === Number(limit))
+                ? (Number(page) * Number(limit)) + 1
+                : ((Number(page) - 1) * Number(limit)) + Number(data.length),
             results: data
         });
     } catch (err) {