+++ src/routes/places/movieTheater.ts
@@ -16,9 +16,12 @@ movieTheaterRouter.get('/search', async (req, res) => {
             endpoint: <string>process.env.API_ENDPOINT,
             auth: req.user.authClient
         });
-        const { totalCount, data } = await placeService.searchMovieTheaters({
-            limit: req.query.limit,
-            page: req.query.page,
+
+        const limit = Number(req.query.limit);
+        const page = Number(req.query.page);
+        const { data } = await placeService.searchMovieTheaters({
+            limit: limit,
+            page: page,
             project: { ids: [req.project.id] },
             name: req.query.name
         });
@@ -52,7 +55,9 @@ movieTheaterRouter.get('/search', async (req, res) => {
 
         res.json({
             success: true,
-            count: totalCount,
+            count: (data.length === Number(limit))
+                ? (Number(page) * Number(limit)) + 1
+                : ((Number(page) - 1) * Number(limit)) + Number(data.length),
             results: results
         });
     } catch (err) {