+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- APIのX-Total-Count削除への対応
+
 ### Deprecated
 
 ### Removed